Refactor / code improvements
